@@ -3,17 +3,31 @@
 source 'https://rubygems.org'
 ruby '2.7.4'
 
-gem 'pg'
-gem 'rails', '~> 5.0.0', '>= 5.0.0.1'
-
+gem 'activerecord-import'
 gem 'bcrypt'
+gem 'bigdecimal', '1.3.5'
+gem 'cancancan'
 gem 'coffee-rails'
+gem 'delayed_job_active_record'
+gem 'devise'
+gem 'exception_notification'
 gem 'jbuilder'
 gem 'jquery-rails'
+gem 'jquery-ui-rails'
+gem 'jwt'
+gem 'pg'
+gem 'pg_search'
+gem 'public_activity'
 gem 'puma'
+gem 'rails', '~> 5.1.0'
 gem 'redis-rails'
+gem 'rolify'
+gem 'roo', '2.5.1'
 gem 'sass-rails'
+gem 'simple_form'
+gem 'twilio-ruby'
 gem 'uglifier'
+gem 'will_paginate'
 
 group :production do
   gem 'newrelic_rpm'
@@ -51,19 +65,3 @@ group :development, :test do
   gem 'shoulda-matchers', '~> 2.0'
   gem 'webdrivers', '~> 4.0'
 end
-
-gem 'activerecord-import'
-gem 'bigdecimal', '1.3.5'
-gem 'cancancan'
-gem 'delayed_job_active_record'
-gem 'devise'
-gem 'exception_notification'
-gem 'jquery-ui-rails'
-gem 'jwt'
-gem 'pg_search'
-gem 'public_activity'
-gem 'rolify'
-gem 'roo', '2.5.1'
-gem 'simple_form'
-gem 'twilio-ruby'
-gem 'will_paginate'
