@@ -3,7 +3,7 @@
 require 'rails_helper'
 require 'capybara'
 
-feature 'e2e Happy Path' do
+describe 'e2e Happy Path' do
   before(:all) do
     @user = FactoryBot.create(:user)
     Capybara.page.current_window.resize_to(1024, 768)
@@ -14,7 +14,7 @@ feature 'e2e Happy Path' do
     expect(page).to have_text('Welcome to Swapsea')
   end
 
-  it 'signs users in and clicks around', :js => true do
+  it 'signs users in and clicks around', js: true do
     visit '/users/sign_in'
     fill_in 'user[email]', with: @user.email
     fill_in 'user[password]', with: 'swapsea'
@@ -52,26 +52,24 @@ feature 'e2e Happy Path' do
     # page.save_and_open_screenshot
 
     # Logout cancelled
-    find("a.dropdown-toggle").click
-    find("a.md-trigger").click
+    find('a.dropdown-toggle').click
+    find('a.md-trigger').click
     # click_link "logout-link"
     # find(link: "a.logout-link").click
 
-
-    click_on "logout-cancel"
+    click_on 'logout-cancel'
     # find(link: "a.logout-cancel").click
 
     # Logout confirmed
-    find("a.dropdown-toggle").click
-    find("a.md-trigger").click
+    find('a.dropdown-toggle').click
+    find('a.md-trigger').click
     # click_link 'logout-link'
     # find("a.logout-link").click
 
     # click_link 'logout-confirm'
     # find("a.logout-confirm").click
-    find("a.btn-success").click
+    find('a.btn-success').click
 
     expect(page).to have_text('Welcome to Swapsea')
   end
-
 end
